@@ -9,19 +9,21 @@ environment:
   sdk: '>=3.3.4 <4.0.0'
 
 dependencies:
+  app_links: ^6.3.3
   app_tracking_transparency: ^2.0.6+1
-  bloc: ^8.1.4
-  bloc_test: ^9.1.7
+  auth0_flutter: ^1.7.2
+  bloc: ^9.0.0
+  bloc_test: ^10.0.0
   cupertino_icons: ^1.0.6
   dart_geohash: ^2.1.0
   dartz: ^0.10.1
   dio: ^5.5.0+1
-  equatable: ^2.0.5
+  equatable: ^2.0.7
   firebase_analytics: ^10.8.0
   firebase_core: ^2.24.2
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.5
+  flutter_bloc: ^9.0.0
   flutter_blue_plus: ^1.32.12
   flutter_clarity:
     git:
@@ -38,7 +40,7 @@ dependencies:
   geodart: ^0.3.6
   geolocator: ^13.0.1
   get_it: ^7.7.0
-  hydrated_bloc: ^9.1.5
+  hydrated_bloc: ^10.0.0
   internet_connection_checker_plus: ^2.5.1
   intl: ^0.19.0
   json_annotation: ^4.9.0
@@ -49,12 +51,12 @@ dependencies:
   sentry_dart_plugin: ^2.1.0
   sentry_flutter: ^8.8.0
   shared_preferences: ^2.3.3
+  single_factor_auth_flutter: ^5.3.0
   socket_io_client: ^3.0.0-beta.2
   solana: ^0.31.0
   syncfusion_flutter_charts: ^27.1.52
   url_launcher: ^6.3.0
   uuid_v4: ^1.0.2
-  web3auth_flutter: ^5.0.4
 
 dev_dependencies:
   build_runner: ^2.4.9
