@@ -2,63 +2,58 @@ import 'package:get_it/get_it.dart' show GetIt;
 import 'package:sky_trade/features/auth/data/data_sources/auth_remote_data_source.dart';
 import 'package:sky_trade/features/auth/data/repositories/auth_repository_implementation.dart';
 import 'package:sky_trade/features/auth/domain/repositories/auth_repository.dart';
-import 'package:sky_trade/features/auth/presentation/blocs/check_sky_trade_user_exists_bloc/check_sky_trade_user_exists_bloc.dart';
-import 'package:sky_trade/features/auth/presentation/blocs/create_sky_trade_user_bloc/create_sky_trade_user_bloc.dart';
-import 'package:sky_trade/features/auth/presentation/blocs/web_3_auth_capture_custom_tabs_closed_bloc/web_3_auth_capture_custom_tabs_closed_bloc.dart'
-    show Web3AuthCaptureCustomTabsClosedBloc;
-import 'package:sky_trade/features/auth/presentation/blocs/web_3_auth_initialization_bloc/web_3_auth_initialization_bloc.dart'
-    show Web3AuthInitializationBloc;
-import 'package:sky_trade/features/auth/presentation/blocs/web_3_auth_login_bloc/web_3_auth_login_bloc.dart'
-    show Web3AuthLoginBloc;
-import 'package:sky_trade/features/auth/presentation/blocs/web_3_auth_logout_bloc/web_3_auth_logout_bloc.dart'
-    show Web3AuthLogoutBloc;
-import 'package:sky_trade/features/auth/presentation/blocs/web_3_auth_register_bloc/web_3_auth_register_bloc.dart'
-    show Web3AuthRegisterBloc;
-import 'package:sky_trade/features/auth/presentation/blocs/web_3_auth_user_session_bloc/web_3_auth_user_session_bloc.dart'
-    show Web3AuthUserSessionBloc;
+import 'package:sky_trade/features/auth/presentation/blocs/auth_0_authentication_bloc/auth_0_authentication_bloc.dart'
+    show Auth0AuthenticationBloc;
+import 'package:sky_trade/features/auth/presentation/blocs/auth_0_logout_bloc/auth_0_logout_bloc.dart'
+    show Auth0LogoutBloc;
+import 'package:sky_trade/features/auth/presentation/blocs/auth_0_user_session_bloc/auth_0_user_session_bloc.dart'
+    show Auth0UserSessionBloc;
+import 'package:sky_trade/features/auth/presentation/blocs/auth_bloc/auth_bloc.dart'
+    show AuthBloc;
+import 'package:sky_trade/features/auth/presentation/blocs/check_sky_trade_user_exists_bloc/check_sky_trade_user_exists_bloc.dart'
+    show CheckSkyTradeUserExistsBloc;
+import 'package:sky_trade/features/auth/presentation/blocs/s_f_a_authentication_bloc/s_f_a_authentication_bloc.dart'
+    show SFAAuthenticationBloc;
+import 'package:sky_trade/features/auth/presentation/blocs/s_f_a_initialization_bloc/s_f_a_initialization_bloc.dart'
+    show SFAInitializationBloc;
 
 final _sl = GetIt.I;
 
 Future<void> registerAuthServices() async {
   _sl
     // BLoCs
+    ..registerFactory<Auth0AuthenticationBloc>(
+      () => Auth0AuthenticationBloc(
+        _sl(),
+      ),
+    )
+    ..registerFactory<Auth0LogoutBloc>(
+      () => Auth0LogoutBloc(
+        _sl(),
+      ),
+    )
+    ..registerFactory<Auth0UserSessionBloc>(
+      () => Auth0UserSessionBloc(
+        _sl(),
+      ),
+    )
+    ..registerFactory<AuthBloc>(
+      () => AuthBloc(
+        _sl(),
+      ),
+    )
     ..registerFactory<CheckSkyTradeUserExistsBloc>(
       () => CheckSkyTradeUserExistsBloc(
         _sl(),
       ),
     )
-    ..registerFactory<CreateSkyTradeUserBloc>(
-      () => CreateSkyTradeUserBloc(
+    ..registerFactory<SFAAuthenticationBloc>(
+      () => SFAAuthenticationBloc(
         _sl(),
       ),
     )
-    ..registerFactory<Web3AuthCaptureCustomTabsClosedBloc>(
-      () => Web3AuthCaptureCustomTabsClosedBloc(
-        _sl(),
-      ),
-    )
-    ..registerFactory<Web3AuthInitializationBloc>(
-      () => Web3AuthInitializationBloc(
-        _sl(),
-      ),
-    )
-    ..registerFactory<Web3AuthLoginBloc>(
-      () => Web3AuthLoginBloc(
-        _sl(),
-      ),
-    )
-    ..registerFactory<Web3AuthLogoutBloc>(
-      () => Web3AuthLogoutBloc(
-        _sl(),
-      ),
-    )
-    ..registerFactory<Web3AuthRegisterBloc>(
-      () => Web3AuthRegisterBloc(
-        _sl(),
-      ),
-    )
-    ..registerFactory<Web3AuthUserSessionBloc>(
-      () => Web3AuthUserSessionBloc(
+    ..registerFactory<SFAInitializationBloc>(
+      () => SFAInitializationBloc(
         _sl(),
       ),
     )
@@ -66,7 +61,9 @@ Future<void> registerAuthServices() async {
     // Repositories
     ..registerLazySingleton<AuthRepository>(
       () => AuthRepositoryImplementation(
-        _sl(),
+        auth0: _sl(),
+        singleFactorAuthentication: _sl(),
+        authRemoteDataSource: _sl(),
       ),
     )
 
