@@ -1,7 +1,7 @@
 const baseRoutePath = '/';
 const loadingRoutePath = '/loading';
-const loginRoutePath = '/login';
-const registerRoutePath = '/register';
+const errorRoutePath = '/error';
+const getStartedRoutePath = '/getStarted';
 const homeRoutePath = '/home';
 const noInternetConnectionRoutePath = '/noInternetConnection';
 const helpRoutePath = '/help';
