@@ -1,28 +1,29 @@
 import 'package:dartz/dartz.dart' show Either, Unit;
 import 'package:sky_trade/core/errors/failures/auth_failure.dart';
-import 'package:sky_trade/core/utils/enums/networking.dart' show AuthProvider;
-import 'package:sky_trade/features/auth/domain/entities/auth_entity.dart';
+import 'package:sky_trade/features/auth/domain/entities/auth_entity.dart'
+    show Auth0UserEntity, SFAUserEntity, SkyTradeUserEntity;
 
 abstract interface class AuthRepository {
-  Future<Either<Web3AuthInitializationFailure, Unit>> initializeWeb3Auth();
+  Future<Either<Auth0AuthenticationFailure, Auth0UserEntity>>
+      authenticateUserWithAuth0();
 
-  Future<Either<Web3AuthAuthenticationFailure, Web3AuthUserEntity>>
-      authenticateWeb3AuthUserUsing({
-    required AuthProvider authProvider,
-    String? credential,
+  Future<bool> checkAuth0UserSessionExists();
+
+  Future<Either<Auth0UserNotFoundFailure, Auth0UserEntity>> get auth0User;
+
+  Future<Either<Auth0LogoutFailure, Unit>> logoutCurrentAuth0User();
+
+  Future<Either<SFAInitializationFailure, Unit>> initializeSFA();
+
+  Future<Either<SFAAuthenticationFailure, SFAUserEntity>>
+      authenticateUserWithSFAUsing({
+    required String? email,
+    required String idToken,
   });
 
   Future<Either<CreateSkyTradeUserFailure, SkyTradeUserEntity>>
-      createSkyTradeUserAnd({
-    required bool subscribeToNewsletter,
-  });
+      createSkyTradeUser();
 
   Future<Either<CheckSkyTradeUserFailure, SkyTradeUserEntity>>
       checkSkyTradeUserExists();
-
-  Future<bool> checkWeb3AuthSessionExists();
-
-  Future<void> captureWhenWeb3AuthCustomTabsClosed();
-
-  Future<Either<Web3AuthLogoutFailure, Unit>> logoutCurrentWeb3AuthUser();
 }
